+++ src/main/java/com/lxp/view/MenuStrategy.java
@@ -4,17 +4,9 @@ import com.lxp.view.command.MenuCommand;
 
 public interface MenuStrategy<T extends MenuCommand> {
 
-	String title();
-
-	String body();
-
-	T[] commands();
+	MenuScreen screen();
 
 	T parse(int input);
 
 	boolean handle(T command);
-
-	default String menuPrefix() {
-		return "";
-	}
 }