+++ src/main/java/com/lxp/view/command/CourseListCommand.java
@@ -10,20 +10,18 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum InstructorCommand implements MenuCommand {
+public enum CourseListCommand implements MenuCommand {
 
-	REGISTER(1, "강사 등록"),
-	LIST(2, "강사 조회"),
-	BACK(3, "뒤로 가기");
+	SELECT(1, "강의 선택"),
+	BACK(2, "뒤로 가기");
 
 	private final int value;
 	private final String label;
 
-	public static InstructorCommand from(int value) {
+	public static CourseListCommand from(int value) {
 		return Arrays.stream(values())
 			.filter(command -> command.value == value)
 			.findFirst()
 			.orElseThrow(() -> new LxpException(ErrorCode.INVALID_INPUT));
 	}
-
 }