+++ src/main/java/com/lxp/view/command/InstructorListCommand.java
@@ -10,23 +10,18 @@ import lombok.RequiredArgsConstructor;
 
 @Getter
 @RequiredArgsConstructor
-public enum ListCommand implements MenuCommand {
+public enum InstructorListCommand implements MenuCommand {
 
-	SELECT(1, "선택"),
+	SELECT(1, "강사 선택"),
 	BACK(2, "뒤로 가기");
 
 	private final int value;
 	private final String label;
 
-	public static ListCommand from(int value) {
+	public static InstructorListCommand from(int value) {
 		return Arrays.stream(values())
 			.filter(command -> command.value == value)
 			.findFirst()
 			.orElseThrow(() -> new LxpException(ErrorCode.INVALID_INPUT));
-	}
-
-	@Override
-	public boolean isSelectable() {
-		return this == SELECT;
 	}
 }